+++ nominal/core/asset.py
@@ -7,6 +7,7 @@ from types import MappingProxyType
 from typing import Iterable, Mapping, Protocol, Sequence, TypeAlias
 
 from nominal_api import (
+    comments_api,
     event,
     scout,
     scout_asset_api,
@@ -56,6 +57,7 @@ class Asset(_DatasetWrapper, HasRid, RefreshableMixin[scout_asset_api.Asset]):
     created_at: IntegralNanosecondsUTC
 
     _clients: _Clients = field(repr=False)
+    created_by_rid: str | None = field(default=None, repr=False)
 
     class _Clients(
         DataSource._Clients,
@@ -69,6 +71,8 @@ class Asset(_DatasetWrapper, HasRid, RefreshableMixin[scout_asset_api.Asset]):
     ):
         @property
         def assets(self) -> scout_assets.AssetService: ...
+        @property
+        def comments(self) -> comments_api.CommentsService: ...
         @property
         def run(self) -> scout.RunService: ...
         @property
@@ -675,6 +679,7 @@ class Asset(_DatasetWrapper, HasRid, RefreshableMixin[scout_asset_api.Asset]):
             labels=tuple(asset.labels),
             created_at=_SecondsNanos.from_flexible(asset.created_at).to_nanoseconds(),
             _clients=clients,
+            created_by_rid=asset.created_by,
         )